updated frontend

+++ backend/controllers/paymentController.js
@@ -27,6 +27,32 @@ const createPayment = expressAsyncHandler(async (req, res) => {
   }
 })
 
+const createRegistrationFee = expressAsyncHandler(async (req, res) => {
+  const {
+    bankAccountName,
+    phoneNumber,
+    transactionId,
+    amount
+  } = req.body
+
+  const payment = await Payment.create({
+    user: req.user._id,
+    bankAccountName,
+    phoneNumber,
+    transactionId,
+    amount,
+    paymentFor: "registration",
+  })
+
+  if (payment) {
+    // res.set('Access-Control-Allow-Origin', 'http://localhost:9000');
+    res.status(201).json(payment)
+  } else {
+    res.status(400)
+    throw new Error('Payment Data')
+  }
+})
+
 // Update user profile
 // route PUT api/user/profile
 // acess Privet
@@ -65,12 +91,43 @@ const getPaymentList = expressAsyncHandler(async (req, res) => {
   res.json({ payments, page, pages: Math.ceil(count / pageSize) })
 })
 
+// @desc Get all users
+// @route GET api/users
+// @acess Privet
+const getPendingPaymentList = expressAsyncHandler(async (req, res) => {
+  const keywords = {
+    status: 'pending',
+  }
+  const pageSize = Number(req.query.pageSize) || 30;
+  const page = Number(req.query.pageNumber) || 1;
+  const count = await Payment.countDocuments({})
+  const payments = await Payment.find({ ...keywords }).sort({ createdAt: 'desc' }).limit(pageSize).skip(pageSize * (page - 1))
+  res.json({ payments, page, pages: Math.ceil(count / pageSize) })
+})
+
+// @desc Get all users
+// @route GET api/users
+// @acess Privet
+const getRegistrationPaymentByUser = expressAsyncHandler(async (req, res) => {
+  const keywords = {}
+  if (req.user._id) {
+    keywords.user = req.user._id
+    keywords.paymentFor = "registration"
+  }else{
+      // res.set('Access-Control-Allow-Origin', 'http://localhost:9000');
+      res.status(404)
+      throw new Error('Please login')
+  }
+  const payments = await Payment.find({ ...keywords })
+  res.json(payments)
+})
+
 // @desc Delete user
 // @route PUT api/users/:id
 // @acess Privet/Admin
 const deletePayment = expressAsyncHandler(async (req, res) => {
   const payment = await Payment.findById(req.params.id)
-  if (payment && payment.status !== 'approved') {
+  if (payment && payment.status !== 'approved' && payment.user === req.user._id) {
     await payment.deleteOne()
     res.json({ message: 'Payment removed' })
   } else {
@@ -84,4 +141,7 @@ export {
   approveOrRejectPayment,
   getPaymentList,
   deletePayment,
+  createRegistrationFee,
+  getRegistrationPaymentByUser,
+  getPendingPaymentList,
 }
